feat: finalize plugin brush tracker implementation
